requirements.txt
torch==1.13.1
torchvision==0.14.1
pillow==9.3.0
pandas==1.5.3
